feat: persist data to local storage
- persist data to local storage
- refactor the file/folder structur
- lint files
- write test for Card, CardList
