+++ src/components/pages/Info/styled.css.tsx
@@ -1,0 +1,16 @@
+import { css } from 'styled-components';
+
+const styles = css`
+  padding: 1rem 5rem;
+
+  .foot {
+    padding: 1rem 0;
+    margin-top: 4rem;
+  }
+
+  @media screen and (max-width: 1000px) {
+    padding: 0 1rem;
+  }
+`;
+
+export default styles;
