+++ src/components/resuableSections/Notes/Update/index.tsx
@@ -2,6 +2,7 @@ import React, {
   FC, useState, useRef, useEffect,
 } from 'react';
 import styled, { StyledComponent } from 'styled-components';
+import styles from './styled.css';
 
 type Props = {
   deleteNote: () => void;
@@ -47,49 +48,7 @@ const Update: FC<Props> & {
 };
 
 Update.Styled = styled.div`
-  width: 100%;
-  border-radius: 5px;
-  flex-flow: column;
-  padding: 1rem;
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-
-  > *:not(:last-child) {
-    margin-bottom: 1rem;
-  }
-
-  > button {
-    width: 2rem;
-    height: 2rem;
-    border-radius: 50%;
-    font-weight: bold;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  textarea {
-    width: 100%;
-    resize: vertical;
-    border-bottom: 1px solid #ced4da;
-    min-height: 10rem;
-    padding: 1rem;
-  }
-
-  & > div:last-child {
-    display: flex;
-    justify-content: center;
-    margin-top: 1rem;
-
-    button {
-      padding: 0.8rem 1.5rem;
-      border-radius: 5px;
-
-      &.del {
-        background: red;
-        margin-left: 3rem;
-      }
-    }
-  }
+  ${styles}
 `;
 
 export default Update;
