+++ src/components/resuableSections/Table/index.tsx
@@ -1,5 +1,6 @@
 import React, { FC } from 'react';
 import styled, { StyledComponent } from 'styled-components';
+import styles from './styled.css';
 
 type Props = { data: Record<string, any> };
 
@@ -17,7 +18,11 @@ const Table: FC<Props> & {
       {Object.entries(data).map(([key, value]) => (
         <tr key={key}>
           <td>
-            <span>{key.replace('_', ' ').toLocaleUpperCase()}</span>
+            <span>
+              {`${key.charAt(0).toUpperCase()}${key
+                .substr(1, key.length)
+                .replaceAll('_', ' ')}`}
+            </span>
           </td>
           <td>
             <span>{value}</span>
@@ -29,40 +34,7 @@ const Table: FC<Props> & {
 );
 
 Table.Styled = styled.table`
-  &,
-  thead,
-  tbody {
-    width: 100%;
-  }
-  thead {
-    background-color: #f8fafd;
-    th {
-      margin: 0;
-      color: #6c757e;
-      white-space: nowrap;
-      border-top: 1px solid #e7eaf3;
-      border-bottom: 3px solid #e7eaf3;
-      padding: 0.625rem;
-      text-align: left;
-    }
-  }
-  tbody {
-    tr {
-      &:hover {
-        background-color: rgba(231, 234, 243, 0.4);
-      }
-      &:not(:last-child) td {
-        border-bottom: 1px solid #e7eaf3;
-      }
-
-      td {
-        white-space: nowrap;
-        padding: 0.625rem;
-        vertical-align: top;
-        max-width: 5em;
-      }
-    }
-  }
+  ${styles}
 `;
 
 export default Table;
