+++ src/components/ui/LoadingSpinner/index.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import styled, { StyledComponent, FlattenSimpleInterpolation, css } from 'styled-components';
-import { ThemeType } from '<hooks>/useTheme';
+import cssStyles from './styled.css';
 
 const LoadingSpinner: FC<Props> & {
   Styled: StyledComponent<'div', any, Props>;
@@ -18,49 +18,14 @@ const LoadingSpinner: FC<Props> & {
 };
 
 LoadingSpinner.Styled = styled.div`
-  ${({ height, positionTop, styles }: Props & { theme: ThemeType }) => `
-    text-align: center;
-    vertical-align: middle;
-    margin: auto;
+  ${({ height, positionTop, styles }: Props) => `
     height: ${height}px;
     top: ${positionTop};
     position: ${positionTop ? 'relative' : 'static'};
 
-
-    div {
-      display: inline-block;
-      width: 25px;
-      height: 25px;
-      border: 3px solid white;
-      border-top: 3px solid #3169cc;
-      border-radius: 50%;
-
-      transition-duration: 1.1s;
-      transition-property: transform;
-      animation-name: rotate;
-      animation-duration: 1.1s;
-      animation-iteration-count: infinite;
-      animation-timing-function: linear;
-
-
-      @keyframes rotate {
-        from {
-          transform: rotate(0deg);
-        }
-        to {
-          transform: rotate(360deg);
-        }
-      }
-    }
-
-    p {
-      display: inline-block;
-      margin-left: 15px;
-      vertical-align: top;
-      line-height: 25px;
-    }
     ${styles}
   `}
+  ${cssStyles}
 `;
 
 type Props = {
