+++ src/helpers/testUtils/Providers.tsx
@@ -1,15 +1,5 @@
 import React from 'react';
-import { ThemeProvider } from 'styled-components';
-import useTheme from '../../hooks/useTheme';
 
-const Providers = ({ children }: { children: JSX.Element }) => {
-  const { theme } = useTheme();
-
-  return (
-    <ThemeProvider theme={theme}>
-      <>{children}</>
-    </ThemeProvider>
-  );
-};
+const Providers = ({ children }: { children: JSX.Element }) => <>{children}</>;
 
 export default Providers as React.ComponentType<{}>;
