--- src/hooks/tests/useTheme.spec.ts
@@ -1,60 +1,0 @@
-import { renderHook, act } from '@testing-library/react-hooks';
-import useTheme, { ThemeType, ThemeMode } from '../useTheme';
-
-describe('useTheme', () => {
-  beforeEach(() => {
-    window.localStorage.removeItem('theme');
-  });
-
-  it('should return theme', () => {
-    const { result } = renderHook(() => useTheme());
-    const { theme } = result.current as { theme: ThemeType };
-
-    expect(theme.mode).toEqual(ThemeMode.Dark);
-  });
-
-  it('should update the theme', () => {
-    const { result } = renderHook(() => useTheme());
-
-    const { theme: initialTheme } = result.current as { theme: ThemeType };
-    expect(initialTheme.mode).toEqual(ThemeMode.Dark);
-
-    act(() => {
-      initialTheme.changeTheme(ThemeMode.Light);
-    });
-
-    const { theme: updatedTheme } = result.current as { theme: ThemeType };
-    expect(updatedTheme.mode).toEqual(ThemeMode.Light);
-
-    act(() => {
-      initialTheme.changeTheme(ThemeMode.Dark);
-    });
-
-    const { theme: anotherUpdatedTheme } = result.current as { theme: ThemeType };
-    expect(anotherUpdatedTheme.mode).toEqual(ThemeMode.Dark);
-  });
-
-  it('should not update the theme', () => {
-    const { result } = renderHook(() => useTheme());
-    const { theme: initialTheme } = result.current as { theme: ThemeType };
-
-    expect(initialTheme.mode).toEqual(ThemeMode.Dark);
-
-    act(() => {
-      initialTheme.changeTheme(ThemeMode.Dark);
-    });
-
-    const { theme: updatedTheme } = result.current as { theme: ThemeType };
-
-    expect(updatedTheme).toEqual(initialTheme);
-  });
-
-  it('should update theme mode from local storage', () => {
-    window.localStorage.setItem('theme', ThemeMode.Light);
-
-    const { result } = renderHook(() => useTheme());
-    const { theme: initialTheme } = result.current as { theme: ThemeType };
-
-    expect(initialTheme.mode).toEqual(ThemeMode.Light);
-  });
-});
